+++ deepen/__init__.py
@@ -1,1 +1,2 @@
+from deepen.activation import *
 from deepen.model import *