+++ deepen/model.py
@@ -1,4 +1,5 @@
 import numpy as np
+from deepen.activation import relu, sigmoid
 
 def initialize_params(layer_dims):
     """Create and initialize the params of an L-layer neural network.
@@ -57,3 +58,43 @@ def linear_forward(A, W, b):
     cache = (A, W, b)
 
     return Z, cache
+
+def layer_forward(A_prev, W, b, activation):
+    """Compute forward propagation for a single layer.
+
+    Parameters
+    ----------
+    A_prev : ndarray
+        Activations from the previous layer of shape (size of previous layer,
+        number of examples).
+    W : ndarray
+        Weights matrix of shape (size of current layer, size of previous layer).
+    b : ndarray
+        Bias vector of shape (size of the current layer, 1).
+    activation : str {"sigmoid", "relu"}
+        Activation function to be used in this layer.
+
+    Returns
+    -------
+    A : ndarray
+        Output of the activation function of shape (size of current layer,
+        number of examples).
+    cache : tuple of (tuple of ndarray, ndarray)
+        Stored for computing the backward pass efficiently.
+
+        linear_cache : tuple of ndarray
+            Stores `cache` returned by `linear_forward()`.
+        activation_cache : ndarray
+            Stores `Z` returned by 'linear_forward()`.
+    """
+
+    Z, linear_cache = linear_forward(A_prev, W, b)
+
+    if activation == "sigmoid":
+        A, activation_cache = sigmoid(Z)
+    elif activation == "relu":
+        A, activation_cache = relu(Z)
+
+    cache = (linear_cache, activation_cache)
+
+    return A, cache